Core/Misc: Fixed some compiler errors, dropped Arena points and Honor points from the characters table

+++ src/server/game/Entities/Object/Updates/UpdateData.cpp
@@ -51,7 +51,7 @@ bool UpdateData::BuildPacket(WorldPacket* packet)
 
     packet->Initialize(SMSG_UPDATE_OBJECT, 2 + 4 + (m_outOfRangeGUIDs.empty() ? 0 : 1 + 4 + 9 * m_outOfRangeGUIDs.size()) + m_data.wpos());
 
-    *packet << uint32(_map);
+    *packet << uint16(_map);
     *packet << uint32(!m_outOfRangeGUIDs.empty() ? m_blockCount + 1 : m_blockCount);
 
     if (!m_outOfRangeGUIDs.empty())
